+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Header Page
 - Footer Page
 - Night mode
+- Scroll to top button
 
 ---